refactor lazy scans -> lazyframes
